requirements.txt
requests~=2.27.1
schedule~=1.1.0
PyYAML~=6.0
urllib3~=1.26.7